simplify a test
